requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Video / image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Plotting / reporting
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0

# Notebook support (optional but useful)
jupyter>=1.0.0
ipykernel>=6.0.0
